task answer

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -10,7 +10,7 @@ int get_bit(unsigned long int n, unsigned int index)
 	int i = 1 << index;
 	int j = (n & i) >> index;
 
-	if (i > n || index >= sizeof(unsigned long int) * 8)
+	if (i > (int)n || index >= sizeof(unsigned long int) * 8)
 	{
 		return (-1);
 	}
